These changes make it possible to tune the maximum number of recursion levels supported when globbing.  A better way of dealing with possible DoS attacks via globbing is via the RLimit* directives, or to disable globbing support altogether by using the UseGlobbing directive.  However, in some cases it may be necessary to support globbing, but to do so in as restrictive a manner as possible.
The default maximum number of levels supported by GNU's globbing is 8.
This patch keeps that default, but allows the administrator, if need be,
to tune that level lower by means of the PR_TUNABLE_GLOBBING_MAX macro,
defined in include/options.h.  When compiling proftpd, it can be defined
to some lower value, e.g.:

  CFLAGS="-DPR_TUNABLE_GLOBBING_MAX=2" ./configure ...