In tracking down an unrelated bug last night, I discovered that, at present, proftpd's parser will gladly handle empty contexts, such as when:
  <Directory blah>
  </Directory>

or

  <Global>
  </Global>

appear in the configuration file.  These empty sets are useless, and serve
only to use up memory unnecessarily; that memory usage is compounded when
directives are merged down into these sets.

This patch checks for such empty contexts (in end_sub_config()), and removes
them from the configuration tree.