Fixing the messages reported when opening the scoreboard (and making all such messages consistent).
Note: we'll need to document, probably in the release announcement, when
the scoreboard format (include/scoreboard.h's PR_SCOREBOARD_VERSION) is
increment.  For standalone-run daemons, there is no problem; for inetd-run
daemons, though, the old scoreboard file will need to be deleted before
the new daemon is started, else the new daemon will complain of the
scoreboard being too old.  (I think this is what we are seeing now with
people upgrading from 1.2.7 to 1.2.8; it's the first time in who knows how
long that the scoreboard version number was increment.)