Fix a minor bug in the reporting of ServerIdent, where the wrong IP address may be displayed.  For example, with the following configuration:
  <VirtualHost 127.0.0.1>
    Bind 1.2.3.4
    ...
  </VirtualHost>

a client connecting to 1.2.3.4 should see "[127.0.0.1]" at the tail end
of the default ServerIdent displayed, rather than the more appropriate
"[1.2.3.4]".  The fix is to use the IP address to which the client connected,
rather than the ServerAddress field of the server_rec to which the client
connected.  As the case with Bind shows, multiple IP addresses may be
associated with the same server_rec, and the IP address displayed to the
connecting client should not "leak" this fact.