Beginnings of support for stackable FSIO layers.  Rather than allowing only a one-to-one mapping between paths and FS objects, this patch allows for one-to-many mappings, linking FS objects together in a linked list for the same path.
At present, only the top layer on the stack, for a given FSIO operation,
will be used.  I anticipate that it will be the responsibility of the called
FSIO handler to decide whether it wants to act as a filter handler, and
pass the parameters farther down the stack, or whether it wants to act
as a source/sink handler and handle the parameters entirely by itself.
This behavior allows for backward compatibility with existing FSIO modules.