Fix an issue noticed by Daniel, using lftp (but not ftp).
When handling an NLST command on an empty directory, lftp would receive
a TCP RST from proftpd, rather than the more polite FIN.  This would happen
because proftpd did a lazy opening of the data connection; the connection
would be opened only when proftpd, when listing the contents of the directory,
encountered something to list.  In the case of an empty directory, there
would be nothing to list.  Hence no accept(2) for the client's data
connection, followed by a close() on the socket.  In the TCP world, this
is "slamming the connection closed", and hence the client would receive
a RST -- the initial handshake was not completed before the socket was
closed.

The fix, then, is trivial: always open the data connection, regardless of
whether there is data to be sent, when handling NLST commands.  (For some
reason this issue does not seem to affect the LIST command.)