Add the ability to display files via filehandles, using pr_display_fh(), in addition to displaying files via paths using pr_display_file().
Why would we want pr_display_fh()?  Many sites, especially those with lots of
users, like DisplayLogin, but they also want to chroot their users.
The handling of DisplayLogin could be changed to open a filehandle prior
to the chroot, if necessary, and thus continue to display a file whilst
allowing chrooted users.  That's the goal, anyway.