Even more build system magic.  With this change, we add support for a new "hint" tag: $Archive$.  This particular tag is *only* honored in a very particular situation: the contrib module uses its own subdirectory (e.g. contrib/mod_foo/) *and* that module is being built as a static module. Contrib modules built as shared modules fall into different linking rules.
The purpose of this $Archive$ tag is to allow the Makefile for that
mod_foo module to be built from multiple source files, which would result
in multiple object files.  All of those object files would need to be
part of the command to link the final `proftpd' executable.  (A shared
module would not need to present for this link step.)

Rather than have multiple object files from contrib modules, it involves
less files (though at the cost of more complexity) to use a single
archive file, e.g. mod_foo.a.  The Makefile for the contrib module, for
its "static" target, will thus need to assemble a mod_foo.a archive from
the various .o files.  When this happens, the contrib module uses the
$Archive$ hint (similar to the $Libraries$ hint) to let proftpd's
module build system know to expect that archive file.