Move a lot of functions, either existing or new, behind PR_USE_DEVEL checks, since a lot of functions are meant only for the debugging purposes of proftpd developers.  (Hopefully this will make the resulting proftpd executable slightly smaller.)  There will be more such debugging functions added; they are meant as hooks utilized by a mod_debug module (again, only for developer use).
The only user-visible change in this is that the SIGUSR2 signal, when
sent to a proftpd process, would cause that process to print out a trace
of its memory pool usage.  Now, a SIGUSR2 signal will cause an event to
be generated -- but the printing of the memory pool usage is gone.  I may
add that back at some point, if/when there is a hue and cry raised.