If an FTP client aborted the data transfer during a directory listing, then a 226 response would be sent *twice*, e.g.:
  226 Transfer complete
  226 Abort successful

would be seen, using the ftp(1) client.  The problem was that mod_ls,
when closing its data connection, did not check to see if the transfer
had been aborted.  Aborted transfers should cause the data connection
to be closed "quietly", i.e. no sending of the "Transfer complete" message.

With this change, mod_ls now checks for an aborted transfer, and will
close the data connection "quietly" when appropriate, thus eliminating the
duplicate 226 responses.