Renamed the Inet "connection" functions to "conn"; no need to have function names that are that verbose.  Updated the attendant callers, and added compatibility macros in compat.h.
Removed older compatibility macros for the log functions from compat.h.

Added handling of the case where fixup_servers() can return -1 after
a restart.  One user was somehow encountering a case where proftpd could
resolve the hostname on startup, but failed to resolve the hostname after
a restart.  This lead to a segfault because the main_server pointer, after
this failure, pointed to old/stale memory.