Added new Session API.  The data for the session are stashed in the session.notes table, but these accessors can perform some intelligent transformation on the stashed value.
Bug#3183 proved that the construction of the pseudo-ttyname, used in
WtmpLog and PAM authentication, should be consolidated, and this has been
done.

Also used the Session API to implement a %{protocol} LogFormat (and
SQLNamedQuery) variable.  This will resolve to "ftp", "ftps", "ssh2",
"sftp", or "scp" as appropriate.  A unit test for regressions of this
feature is included.