When handling the LIST output, the outputfiles() function used to do 4 sorts, linearly scanning the list of files built up using addfile(). Ostensibly this list is used for sorting 1) a la -t or -S, and 2) for supporting -C.  These 4 sorts are needed only for -C, which is not used that often.  Thus it is now conditional.
In the future, the entire use of the struct filename structure (and the
allocations out of the fpool pool which need to be freed after all of the
sorting) can be avoided for every case _except_ when -C, -S, or -t is used.
This too can add some performance gains to the perceived LIST response time.