Bug#3225 - Opening of files during restart can inappropriately use stdin/stdout/stderr descriptors, leading to bad behavior.
This patch adds a new pr_fsio_get_usable_fd() function, for obtaining an
fd (via dup(2)) which is NOT stdin/stdout/stderr.  It is to be used by
functions which can be called in the daemon process during startup/restart
to ensure that the obtained descriptor will not cause other issues (i.e.
with other code which assumes the use of stdin/stdout/stderr).  I suspect
that this has been the underlying cause of several inexplicable crashes/
log messages during restarts.