The mod_sftp module was not honoring a configuration like:
  <Limit WRITE>
    DenyAll
  </Limit>

  <Directory ~/*/*>
    <Limit STOR>
      AllowAll
    </Limit>
  </Directory>

the same way between SFTP and SCP uploads.  SFTP uploads were allowed properly,
but SCP uploads were not.  Turned out to be a problem with the path that
was being used in the dir_check() calls in the SCP code; the full path was
not being used as it should.

Also fixed a problem where an SCP client might send duplicate filename info
(once in the channel request, once in the fileinfo control message).