In the mod_sftp unit test for handling SIGTERM, I noticed that one of the configured SFTPHostKey files kept disappearing (as in, being deleted) once the server was shut down.  Another mod_sftp user has reported seeing the same symptoms, only intermittently.  I was finally able to track down the cause.
Turns that the PidFile handling code was not being very careful about making
sure that the static pidfile_path pointer kept pointing to the same memory.
The pointer used for looking up the PidFile directive was the same pointer;
and if the lookup failed/changed, the pointer could change.  And, say, point
at some other file.  Then, when the server shut down and deleted its PidFile,
well, that other file might be deleted.  Not desirable behavior.