Fix a rather egregious bug in the SCP download code.  We were reading the entire file being downloaded into memory, and sending (or buffering) the data for it BEFORE returning back to handling any messages which may have been sent by the client while we were reading.  If the file is larger than the default/initial window size, then we end up buffering.  We only drain that buffered data once the file had been read in its entirety, and we go back and handle any WINDOW_ADJUST messages, which then open the window again and we can drain our data.
Now imagine if that file is large, say, >4GB.  The client sends its
WINDOW_ADJUST messages, but mod_sftp never gets around to listening to them
for quite a while.  Not good.  The fix, then, is to have the SCP download
code check the window size periodically.  If the window closes, then handle
enough messages from the client until the window opens again.