Bug#3229 - LangDefault does not accept any setting on FreeBSD.
There were multiple issues involved in this one.  First, mod_lang needed to
handle setlocale(3) where the translated files shipped with proftpd (which
do not have ".UTF-8" in their names) did not match the languages accepted by
setlocale() on the platform (e.g. where setlocale() expected names which *did*
end in ".UTF-8"); this is handled by populating a table of language aliases.

Second, the initial setlocale(LC_ALL, "") call, to pick up/honor the various
environment variables as one would expect from a Unix program, was happening
in quite a few places in mod_lang.  Not good.  So this was moved into the
main() function, as part of the initial startup.  And yes, it does check that
the setlocale() function call succeeds, and warns about badly configured
environment variables if not.

Last, we had to always set the LANG environment variable, in order to ensure
that gettext used the translated files of the language to which we had
switched.  While the old code would work on Linux (and on Solaris...I think),
it would NOT work on FreeBSD (and possibly other BSD-related platforms).  It
was a long arduous slog to finding that using LANG was required.  Ugh.