Revisit the fix put in place for Bug#2367.  First, add the test case to the LIST testsuite, for regressions.  Then look at a different fix, a simpler one: remove the GLOB_PERIOD flag from the flags used when globbing.  This causes the test case to pass.
This is done in order to make possible another case: list a directory whose
name legitimately contains a glob character ('[' in the case of one of the
users in the online forums).  The previous fix for Bug#2367 prevented the
proper listing of such a directory.  The newer, simpler fix handles this
legitimate directory name properly, while still preserving behavior which
matches that of the shell.