Fix up some issues found in mod_tls when running its test cases.
First, as per Bug#3325, make the TLSRequired/AllowPerUser check a little
smarter, so that the published AllowPerUser example in the TLS howto works
properly, but the check still retains its validity.

When checking for client-initiated renegotiations, make sure we don't punish
the client for *new* sessions for *data* connections.

Make sure that "TLSRequired auth" is enforced properly if the client is NOT
using SSL/TLS.

Include MLSD in the list of commands which can trigger data transfers, and thus
fall under the "TLSRequired data" checks.