Fix a problem in mod_shaper, noted in this forums post:
  http://forums.proftpd.org/smf/index.php/topic,1628.0.html

where the queue used by mod_shaper could become "clogged", full of old
stale messages.  The fix, as noted as well, is to clear the queue of any
messages for the destination process before sending a new message to that
process.

Added a regression test to reproduce the behavior; it required a manual
change to the mod_shaper code in order to fill up the queue in the time of
the test.  Good enough for now.