Bug#3387 - SIGHUP will eventually cause certain modules (as DSO modules) to segfault.
Register timers in the daemon process via a 'core.postparse' listener rather
than a 'core.startup' listenver, and remove the registered timers in
a 'core.restart' listener.  Otherwise, the previous registered timer can
point to invalid memory if the registering module is a DSO module; SIGHUPs
cause DSO modules to be unloaded/reloaded.