Bug#3418 - %U sometimes showing up as "(none)" in ExtendedLog.
When removing an entry from the linked list "chain" in a table, the table
code was improperly setting the head of the chain to the removed entry's
next pointer (e->next) when it shouldn't have been.  In this case, the
removed entry was the last of two entries in the list; its removal was causing
the head of the chain to be set to null.