Bug#3445 - Support SFTPClientMatch option to enable pessimistic NEWKEYS behavior.
This changes the default behavior, such that mod_sftp always sends its
NEWKEYS message to the client before reading the client's NEWKEYS message;
this matches the OpenSSH behavior (and is what many clients want).  The
SFTPClientMatch option added previously for this bug was changed to
'pessimisticNewkeys', to be used if there are any clients which expect the
old mod_sftp behavior.