Bug#3452 - mod_sftp does not advertise its supported SFTP extensions for protocol version 5.
I read through the WinSCP source code to find out how it parsed the
badly-defined extensions field of the "supported" extension, in the protocol
version 5 draft.  The "supported2" extension, for protocol version 6, is
at least nice enough to have a separate field of the extensions count, where
the older "supported" does not.  Whee.