revert r1.83 in favor of mod_ldap bzr r98
we shouldn't let user lookup failure prevent us from looking up any
supplemental groups. this is primarily so mod_ldap can be used to look
up group membership even when the user wasn't authenticated with mod_ldap.

this feels like a corner use-case, but i don't see a compelling reason
we shouldn't support it since the change to prevent the segfault in
this case is straightforward.