Generate a 'core.shutdown' event when the daemon is shutting down.
Currently, a 'core.exit' event is generated when the daemon shuts down.
However, this event is also used to indicate when a session is closing; this
forces modules to register two different event handlers for the same event
name, depending on the context.  It is far cleaner (and less bug-prone) to
have two distinct events for the two different types of "end of processing"
behavior.

Gradually, the generation of the 'core.exit' event at daemon shutdown time
will be phased out.