Add some guards to mod_delay, to try to curb overly excessive delays.
First, initialize the delay value records to use values of -1, and then make
sure that such negative values are skipped when building the list of values
from which to select the median.

Check that the selected median is greater than or equal to zero before using
it.

Last, add an absolute limit on the selected delay; currently set to 15 minutes.