Bug#3526 - CPU usage at 100% when checking HideFiles pattern.
The build_dyn_config() function was using different names for the .ftpaccess
config_recs it created versus the ones it looked up.  This lead to an
explosing in the number of config_recs created, which in turn lead to
huge amounts of memory and CPU being taken up.  Thanks to Thomas Shinnick
for diligently tracking down the problem, and for providing a thoroughly-
researched fix.