Added new Command API functions:
  pr_cmd_get_id()
  pr_cmd_cmp()

These functions are used to assign an ID to a cmd_rec, identifying the FTP
command of that cmd_rec.  Code which wants to check which FTP command a cmd_rec
is can then use pr_cmd_cmp() and compare using a numeric ID, rather than
calling strcmp(3) repeatedly on the same cmd_rec throughout the code.

This is done with an eye toward reducing wasteful/unnecessary strcmp(3) calls
throughout the proftpd code base.