Fix some issues with ProFTPD'd pr_table_t handling.
First, the default hashing function was fixed so that it actually used the
different values in the given key.  Previously, it only ever used the first
value.

Second, randomize the hash output by generating a random "seed" value on table
creation time.  Any custom key hash callbacks need to be updated to accept
this seed value as one of the hashing inputs.  This will help to mitigate
any attacks which seek to force the table API into the worst-case performance
scenario, where all keys are effectively hashed to the same chain, and thus
linear scans are done, burning CPU time.

Last, add a limit to the maximum number of items that a table can hold.  This,
too, helps to mitigate any attacks/abuses that can be done on tables.  This
maximum limit is 8K by default, and can be modified by a call to pr_table_ctl.