Bug#3707 - Add request/transfer ID to the logging of the initial and closing commands for SFTP file transfers.
This adds a "sftp.file-handle" note to the cmd->notes table for SFTP
transfer-related requests.  This note can then be logged in a custom LogFormat
using the new support for notes, e.g. '%{note:sftp.file-handle}'.

Note that I also fixed the logging of %f LogFormat variables for SFTP CLOSE
requests; the session.xfer struct was being cleared by the RETR/STOR handles
to which the CLOSE request handler dispatched, and thus when the CLOSE request
was dispatched, mod_log had no session.xfer information to work with.  Now
fixed.