Add more information to the cmd_rec logged for a WRITE request, such that the %r LogFormat variable will also include the file offset and chunk length fields from the WRITE request.
Make sure to reset the cmd_id member for cmd_recs when changing the cmd_rec
name; otherwise, the cmd_id value will not be reset, and logging may not Do
The Right Thing(tm) with that cmd_rec.  This was the cause of %f LogFormat
variables not being interpolated properly by mod_log for WRITE requests.