Fix the handling of event notification subscription.  Specifically, modules that were loaded last would receive the event notification last.
And if that notification were for the 'core.restart' event, AND the subscribing
module was built as a shared module, this would mean that that module would
NEVER be notified; the mod_dso 'core.restart' event handler would be notified
first, and would unload all shared modules.

The fix, then, is to have the modules loaded last have their notifications be
delivered first.  This was the intended ordering in the first place.