Fix an interesting bug only encountered if the following config is used:
  DefaultAddress <ipv6-addr>
  UseIPv6 off

That is, the server is configured with an IPv6 address, and then later, in
the config, support for IPv6 is disabled.  Without this patch, a strange
segfault will ensue on startup.

Note that if the order of the directives is reversed, so that UseIPv6 occurs
before the DefaultAddress directive, then startup fails as one might expect,
since the IPv6 address won't be resolved unless IPv6 support is enabled.