Configure: Refactor module-selection code This is a big patch, but it most of it is moving code around to facilitate proper and simplified handling of selected modules.
Several modules now get added to the SHARED_MODULE_OBJS that were
erronously not added to that list before, they were still compiled
due to other factors pulling them in it seems.

This also makes the modules listing at the end of configure also
contain core modules and a few modules that were invisible due to
the above bug in the code.