Bug#4166 - mod_sftp sessions consume large amounts of memory due to rekeying.
In order to debug this, I had to add code to mod_sftp which handles the
SIGUSR2 signal by dumping the memory pools of the process.  This, in turn,
required some slight adjustments to a similar handling of SIGUSR2 that mod_xfer
does, but which was only needed if mod_shaper is present.