Properly handle multiple TLSDHParamFiles, if configured.
Adjust the DH parameter length to be the size of the RSA/DSA private key, for
greater strength, rather than using the given length (which will only ever
be 512 or 1024 bits, even for 2048-bit server certs/keys).

If OpenSSL-1.0.2 or later are used, automatically negotiate the best EC,
rather than always falling back to the prime256v1 group.