Make sure that the port number logged in the "RootRevoke in effect" messages is accurate.
Enforce the case where RootRevoke is on, but the client has ignored our
error responses to PORT/EPRT, and had requested a data transfer anyway.
We do this by returning an error response for the data transfer command
(e.g. LIST, STOR, etc) before we even attempt to make the connection.