Bug#4235: Recursive SCP uploads of directories fail with "No such file or directory".
The SCP handling code was not clearing the statcache when it should.  The
statcache clearing code did not properly handle relative paths, even when
the clearing function WAS called.  And the pr_fsio_smkdir() function, after
creating the directory securely, did not properly clear the statcache for
the created directory.