Bug#4299: TimeoutLogin no longer working for SFTP connections.
The root cause was the blocking of alarms when reading/writing SFTP packets;
this was done to avoid corrupting packets during rekeying.  But it had the
unwanted side effect of blocking alarms like TimeoutLogin.  To balance the
TimeoutLogin need, we block alarms for packet IO only after the client has
authenticated.