Upon further reflection, I think the blocking of alarms when _reading_ SSH2 packets is too intrusive, and (importantly) not really needed for addressing Bug#4216.
Reading SSH2 packets reads the components into a local object, thus there is
not the same chance of interruptions as for writing a packet.  (And this
admittedly makes me wonder whether a redesign of the packet writing routines,
to use a local object, might be a better long-term solution to Bug#4216.)

Noticed that the bug number listed in NEWS was wrong; fixed that.