Fix false-positives in AIX function detection
Tested on Solaris 10/SPARC. Without this change the configure script
wrongly detects the authenticate() and friends AIX functions, yielding
scary compile errors later on.

The reason is that the configure test programs protect the usersec.h
with a CPP guard. On Solaris 10, the condition is false, thus the header
(that is not present on Solaris) is not included such that the following
function call doesn't yield a compile error because the function is
implicitly declared.