- let proftpd to drop core, when it dies because of segfault
The Stack Trace we get from customers is not that helpful sometimes.
We would like our customers to send us also a core file from dying
process. Unfortunately the proftpd is not allwed to drop the core.
Not sure if it is a bug or intention, hence I'm sending a patch,
which prevents proftpd dying on SigSegv to call exit(3C).