004.proftpd-retry.patch
FTP server in active mode may hit EADDRINUSE, when
connecting back to FTP client, when client port
range is limited. There is no easy way to quantify
'limited port range'. In some cases 100 ports on NAT
pool may be well enough (when files are long and
transmission lines are slow) in some cases 100 ports
in pool is not sufficient, when lines are fast and
files transferred are short.

As soon as file is transferred the connection is
closed (via shutdown(2)). This transfers socket/session
to TIMEWAIT state, which has typically 30sec duration.
However this seems to vary system to system [1].

With short files and fast lines 100 ports may get
wasted in couple of seconds (when doing mget), preventing
server to establish new connections. When server
is connecting to port number, which is still kept
in timewait state, the connect(2) operation fails
with EADDRINUSE.

Besides using this patch there are two ways
to fix/workaround the situation:
   - either let server to use any source port
     not just port 20

   - or set REUSEADDR/REUSEPORT socket option,
     so server will be connect to destination
     although the previous session is still
     in TIMEWAIT state.

The proposed patch works reasonably well, but
I can imagine it deserves some love, like introducing
yet another configuration to option to make
number of retry attempts tunable.

[1] https://en.wikipedia.org/wiki/Maximum_segment_lifetime