Issue #750: Fix mod_digest's handling of the `POST_CMD` phase for the `PASS` command.
The command table entry inadvertently had the "requires authentication" flag
set to `TRUE`, which does not make sense for a `POST_CMD` phase handler for
the `PASS` command, at least in the context of an SFTP login.  This would cause
`mod_digest` to effectively short-circuit the further dispatching of the `PASS`
command for that phase.  Due to module load ordering, this means that the
`mod_auth` module would not be called, and thus would not be able to enforce
limits.