mod_sftp: Fix check whether st_mode change is necessary
When creating a file, the mode has to be sent with the fxp open call.
This doesn't contain the file type bits, but it is later compared against
the full st_mode which contains them. Therefore it attempts to do an
unnecessary chmod, which can fail due to SITE_CHMOD being disabled.