Specify file mode for open with O_CREAT or O_TMPFILE
Use of open() with O_CREAT or O_TMPFILE in second argument requires a
third argument to specify the required file mode. Without this value,
a random value from the stack is used instead.

Furthermore, with glibc and compiling with -D_FORTIFY_SOURCE=1 or
higher, the absence of the third argument causes errors like this:

In file included from /usr/include/fcntl.h:329,
                 from /usr/include/sys/file.h:24,
                 from ../include/os.h:90,
                 from ../include/conf.h:32,
                 from api/tests.h:30,
                 from api/scoreboard.c:27:
In function 'open',
    inlined from 'scoreboard_set_test_fn' at api/scoreboard.c:146:8:
/usr/include/bits/fcntl2.h:50:11: error: call to '__open_missing_mode' declared with attribute error: open with O_CREAT or O_TMPFILE in second argument needs 3 arguments
   50 |           __open_missing_mode ();
      |           ^~~~~~~~~~~~~~~~~~~~~~