Match assert format strings and their parameters
Mostly this is about handling objects of type 'size_t' by using the
'%lu' format and casting the object to 'unsigned long'.

If we could assume a C99 compiler then we could just use the '%zu'
format and not bother with the cast.

There are also a few other miscellaneous mismatches fixed.