Issue #1597: Properly handle `From` directives inside a `<Class>` section that is itself within a `<Global>` section.
The underlying cause was the lack of creating a "Class" config context on
the parser stack for the entire `<Class>` section.  Without this, the
`From` directives were not being associated into the `CONF_CLASS` configuration
context, which in turn meant that the directive handler checks for proper
context were failing unexpectedly.