Issue #1617: The SSH, SFTP `ExtendedLog` classes have been broken for a while, due to filtering in the Jot API of commands lacking known command IDs.
The core ProFTPD engine has command IDs for FTP commands, but not for the
SSH, SFTP requests supported by the `mod_sftp` module.  The workaround is
for `mod_sftp` to manually assign a fake command ID, such that the Jot API
will think these commands are known, and thus handle them as expected.