When looking into buffer sizes, and handling of buffers, during downloads using TLS, I noticed that we _always_ did a `memcpy` of the given buffer (as allocated in mod_xfer, and filled from disk) into a different buffer.
This `memcpy` is necessary if we are performing ASCII translation on that data, yes -- but *not* necessary if we are *not* performing ASCII translation.  Consider the case of a download of a large binary file.  We read a large chunk of the file data from disk into a buffer, then _copy that buffer_, then write out the same data (from the second buffer) to the network.  Why?

This, then, is an experiment to attempt avoiding an unnecessary `memcpy` (and its latency) during downloads that do not need the intermediate buffer.