Don't include <openssl/engine.h> if ENGINE support not detected (#1816)
OpenSSL ENGINE support is deprecated in Fedora 41 onwards:
https://fedoraproject.org/wiki/Changes/OpensslDeprecateEngine

As part of this change, OpenSSL is still built with ENGINE support in
order to maintain compatibility with existing applications, but the
<openssl/engine.h> header is moved to a different package so that
users have to make an explicit choice to support it.

Since proftpd already supports building without ENGINE support,
we can qualify inclusion of <openssl/engine.h> using the results
of the configure check. If there's no ENGINE support, there's no
point including it.