mod_sftp: Fix check whether st_mode change is necessary (#1033)
When creating a file, the mode can be sent with the SFTP `OPEN` request.
This doesn't contain the file type bits, but it is later compared against
the full `st_mode` which contains them. Therefore it attempts to do an
unnecessary chmod, which can fail due to `SITE_CHMOD` being possibly disabled via configuration.