use $(LIBTOOL) to install proftpd
With slibtool 'make install' will install the slibtool wrapper script
instead of the actual executable file. This is because with slibtool
the executable is compiled into the .libs directory while the file in
the root directory is a wrapper script where with GNU libtool the
executable will be placed into the root directory. Using $(LIBTOOL) to
install the executable will ensure that both implementations install the
correct file.

Gentoo-Issue: https://bugs.gentoo.org/953968